fix custom ids
There were three types of fixes depending on the situation.

1. Remove custom id.  Each heading has a default id that comes from
transforming the header contents into spinal-case.  In many cases, the
custom id matched the auto-id and in these cases it was removed.
2. In (one or more) cases, the custom id was more appropriate to a subsection and
it was therefore moved to that subsection.
3. Where neither 1 nor 2 worked, an anchor tag was placed just before
the header with the id of interest.  This makes both the auto-id and the
custom id available and makes for a helpful link in both cases.